consolidate CreateDirectory and clear readonly flag
